ci(client): add workflow to deploy Storybook to GitHub Pages
This commit introduces a new GitHub Actions workflow to automatically deploy Storybook to GitHub Pages when changes are pushed to the 'develop' branch or when pull requests target the 'develop' branch. Additionally, the `.gitignore` file is updated to ignore the `storybook-static` directory, and a new script `build-storybook` is added to the `package.json` to build Storybook for production.